requirements.txt
python-telegram-bot
flask
python-telegram-bot[job-queue]
mysql-connector-python
python-dotenv
firebase-admin
